ServiciosDeAreaYCustomControllerBase

+++ ProyectCRM.Data/Repositories/AreaRepository.cs
@@ -6,9 +6,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ProyectCRM.Data
+namespace ProyectCRM.Data.Repositories
 {
-    public class AreaRepository : IRepository<Area>
+    public class AreaRepository : IRepositoryBase<Area>
     {
         private readonly AppDbContext _context;
 
@@ -24,12 +24,22 @@ namespace ProyectCRM.Data
             return area;
         }
 
-        public async Task DeleteAsync(int id)
+        public async Task<bool> DeleteAsync(Guid id)
         {
             var area = _context.Area.FirstOrDefault(a => a.id == id);
+            if (area == null)
+            {
+                return false; // Area not found
+            }
             _context.Area.Remove(area);
             await _context.SaveChangesAsync();
+            return true;
 
+        }
+
+        public Task<bool> FiltrarPorNombreAsync(string args)
+        {
+            return _context.Area.AnyAsync(a => a.nombre == args);
         }
 
         public async Task<IEnumerable<Area>> GetAllAsync()
@@ -38,21 +48,25 @@ namespace ProyectCRM.Data
 
         }
 
-        public async Task<Area> GetByIdAsync(int id)
+        public async Task<Area> GetByIdAsync(Guid id)
         {
             return await _context.Area.FirstOrDefaultAsync(a => a.id == id);
         }
 
-        public async Task<Area> UpdateAsync(int id, Area areaUpdate)
+        public async Task<Area> UpdateAsync(Guid id, Area areaUpdate)
         {
             var area = await _context.Area.FirstOrDefaultAsync(a => a.id == id);
-
+            if(area == null)
+            {
+                throw new Exception("La entidad no existe");
+            }
             area.id = id;
             area.nombre = areaUpdate.nombre;
             _context.Area.Update(area);
             await _context.SaveChangesAsync();
             return area;
         }
+
     }
 
 }
